# docs/swagger.yaml
basePath: /api/v1
definitions:
  cmd_api.CreateEventRequest:
    properties:
      category:
        type: string
      description:
        type: string
      end_time:
        type: string
      location:
        type: string
      max_attendees:
        type: integer
      start_time:
        type: string
      ticket_tiers:
        items:
          $ref: '#/definitions/cmd_api.TicketTierReq'
        minItems: 1
        type: array
      title:
        type: string
    required:
    - category
    - description
    - end_time
    - location
    - max_attendees
    - start_time
    - ticket_tiers
    - title
    type: object
  cmd_api.CreateOrderRequest:
    properties:
      reservation_id:
        type: string
    required:
    - reservation_id
    type: object
  cmd_api.EventResponse:
    properties:
      category:
        $ref: '#/definitions/eventix-api_internal_models.EventCategory'
      created_at:
        type: string
      description:
        type: string
      end_time:
        type: string
      id:
        type: string
      location:
        type: string
      max_attendees:
        type: integer
      organizer_id:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/eventix-api_internal_models.EventStatus'
      ticket_tiers:
        items:
          $ref: '#/definitions/cmd_api.TicketTierResponse'
        type: array
      tickets_sold:
        type: integer
      title:
        type: string
    type: object
  cmd_api.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  cmd_api.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/eventix-api_internal_models.OrderStatus'
      ticket_count:
        type: integer
      total_amount:
        type: number
    type: object
  cmd_api.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  cmd_api.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  cmd_api.ReserveTicketRequest:
    properties:
      quantity:
        maximum: 10
        minimum: 1
        type: integer
      tier_id:
        type: string
    required:
    - quantity
    - tier_id
    type: object
  cmd_api.TicketResponse:
    properties:
      created_at:
        type: string
      event_id:
        type: string
      event_title:
        type: string
      id:
        type: string
      qr_code:
        type: string
      status:
        $ref: '#/definitions/eventix-api_internal_models.TicketStatus'
      tier_name:
        type: string
    type: object
  cmd_api.TicketTierReq:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      quantity:
        minimum: 1
        type: integer
    required:
    - name
    - price
    - quantity
    type: object
  cmd_api.TicketTierResponse:
    properties:
      available:
        type: integer
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
      sold:
        type: integer
    type: object
  cmd_api.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  cmd_api.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  cmd_api.ValidateQRRequest:
    properties:
      event_id:
        type: string
      qr_code:
        type: string
    required:
    - event_id
    - qr_code
    type: object
  cmd_api.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  eventix-api_internal_models.EventCategory:
    enum:
    - music
    - sports
    - arts
    - technology
    - business
    - education
    - other
    type: string
    x-enum-varnames:
    - CategoryMusic
    - CategorySports
    - CategoryArts
    - CategoryTechnology
    - CategoryBusiness
    - CategoryEducation
    - CategoryOther
  eventix-api_internal_models.EventStatus:
    enum:
    - draft
    - under_review
    - published
    - active
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - EventDraft
    - EventUnderReview
    - EventPublished
    - EventActive
    - EventCompleted
    - EventCancelled
  eventix-api_internal_models.OrderStatus:
    enum:
    - pending
    - paid
    - failed
    - cancelled
    - refunded
    type: string
    x-enum-varnames:
    - OrderPending
    - OrderPaid
    - OrderFailed
    - OrderCancelled
    - OrderRefunded
  eventix-api_internal_models.TicketStatus:
    enum:
    - reserved
    - active
    - used
    - cancelled
    - refunded
    type: string
    x-enum-varnames:
    - TicketReserved
    - TicketActive
    - TicketUsed
    - TicketCancelled
    - TicketRefunded
host: localhost:8080
info:
  contact:
    email: support@eventix.com
    name: API Support
  description: Production-grade ticket booking system API for events, organizers,
    and attendees
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Eventix Ticket Booking API
  version: "1.0"
paths:
  /admin/stats:
    get:
      consumes:
      - application/json
      description: Get platform statistics and metrics (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Get admin statistics
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/cmd_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/cmd_api.TokenResponse'
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      summary: User login
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
                  token_type:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/cmd_api.UserResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      summary: Register a new user
      tags:
      - Auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verify a user's email address using verification token
      parameters:
      - description: Verification token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      summary: Verify email address
      tags:
      - Auth
  /checkin/validate:
    post:
      consumes:
      - application/json
      description: Validate a ticket QR code for event check-in (Organizer/Admin only)
      parameters:
      - description: QR validation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api.ValidateQRRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Validate QR code
      tags:
      - Check-in
  /events:
    get:
      consumes:
      - application/json
      description: Get a list of all published events
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/cmd_api.EventResponse'
                type: array
              pagination:
                properties:
                  limit:
                    type: integer
                  page:
                    type: integer
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
      summary: List all events
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new event (Organizer/Admin only)
      parameters:
      - description: Event details
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/cmd_api.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/cmd_api.EventResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Create a new event
      tags:
      - Events
  /events/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/cmd_api.EventResponse'
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      summary: Get event by ID
      tags:
      - Events
  /orders:
    post:
      consumes:
      - application/json
      description: Create an order for reserved tickets
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/cmd_api.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/cmd_api.OrderResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Create an order
      tags:
      - Orders
  /orders/my-orders:
    get:
      consumes:
      - application/json
      description: Get all orders placed by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/cmd_api.OrderResponse'
                type: array
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Get user's orders
      tags:
      - Orders
  /tickets/my-tickets:
    get:
      consumes:
      - application/json
      description: Get all tickets owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/cmd_api.TicketResponse'
                type: array
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Get user's tickets
      tags:
      - Tickets
  /tickets/reserve:
    post:
      consumes:
      - application/json
      description: Reserve a ticket for an event (15-minute hold)
      parameters:
      - description: Ticket reservation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cmd_api.ReserveTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  expires_at:
                    type: string
                  reservation_id:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Reserve a ticket
      tags:
      - Tickets
  /users/me:
    get:
      consumes:
      - application/json
      description: Get authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/cmd_api.UserResponse'
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                properties:
                  code:
                    type: string
                  message:
                    type: string
                type: object
              success:
                type: boolean
            type: object
      security:
      - OAuth2Password: []
      summary: Get current user profile
      tags:
      - Users
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      read: Grants read access
      write: Grants write access
    tokenUrl: /api/v1/auth/login
    type: oauth2
swagger: "2.0"